Move user sessions types to /models
