+++ src/functions/auth/get-one-user-active-sessions.ts
@@ -6,6 +6,7 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
+import { RawUserSessions, toUserSessions, UserSessions } from '../../models';
 import {
 	APIFunctionMakerOptions,
 	buildHTTPRequest,
@@ -27,44 +28,11 @@ export const makeGetOneUserActiveSessions = (makerOptions: APIFunctionMakerOptio
 
 		try {
 			const raw = await fetch(url, { ...req, method: 'GET' });
-			const data = await parseJSONResponse<RawSuccessResponse>(raw);
-			return beautifyResponse(data);
+			const data = await parseJSONResponse<RawUserSessions>(raw);
+			return toUserSessions(data);
 		} catch (err) {
 			if (err instanceof Error) throw err;
 			throw Error('Unknown error');
 		}
 	};
 };
-
-interface RawSuccessResponse {
-	UID: number;
-	User: string;
-	Sessions: Array<{
-		Origin: string; // IP eg. "127.0.0.1"
-		LastHit: string; // Timestamp eg. "2020-08-04T12:55:40.6312945Z"
-		TempSession: boolean;
-		Synced: boolean;
-	}>;
-}
-
-export interface UserSessions {
-	userID: string;
-	username: string;
-	sessions: Array<{
-		origin: string;
-		lastHit: string;
-		isTemporary: boolean;
-		isSynced: boolean;
-	}>;
-}
-
-const beautifyResponse = (raw: RawSuccessResponse): UserSessions => ({
-	userID: raw.UID.toString(),
-	username: raw.User,
-	sessions: raw.Sessions.map(s => ({
-		origin: s.Origin,
-		lastHit: s.LastHit,
-		isTemporary: s.TempSession,
-		isSynced: s.Synced,
-	})),
-});
