+++ src/models/actionable/raw-updatable-actionable.ts
@@ -1,3 +1,11 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
 import { RawNumericID, RawUUID } from '../../value-objects';
 import { RawActionableAction, RawActionableTrigger } from './raw-actionable';
