+++ src/models/actionable/updatable-actionable.ts
@@ -1,3 +1,11 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
 import { ConditionalPartial, PartialProps } from '../../functions/utils';
 import { NumericID, UUID } from '../../value-objects';
 import { ActionableAction, ActionableTrigger } from './actionable';
